refactor: вынести фильтр в отдельную функцию
