+++ src/components/SavedMovies/SavedMovies.js
@@ -5,7 +5,7 @@ import Header from '../Header/Header';
 import SearchForm from '../SearchForm/SearchForm';
 import MoviesCardList from '../MoviesCardList/MoviesCardList';
 import Footer from '../Footer/Footer';
-import { filter } from '../../utils/constants';
+import filter from '../../utils/filter';
 
 function SavedMovies({
 	menuActive,
